+++ ambassador-auth/ambassador_auth/app.py
@@ -1,64 +1,123 @@
-import config as cfg
 import logging
-import requests
+import json
+import os
 
-from flask import Flask, request, Response, make_response
-from flask_caching import Cache
+from flask import Flask, request, Response, make_response, json
+from jose import jwt
+from jose.exceptions import JWTError, ExpiredSignatureError, JWTClaimsError
 from urllib.parse import unquote_plus
 
-config = {
-    "DEBUG": False,
-    "CACHE_TYPE": "filesystem",
-    "CACHE_THRESHOLD": 600,  # max number of items the cache stores before it starts deleting values
-    "CACHE_DEFAULT_TIMEOUT": 86400,  # cache will last for one day
-    "CACHE_DIR": "./cache/"
-}
+config = json.load(open("config.json"))
+
 app = Flask(__name__)
 
 app.config.from_mapping(config)
-cache = Cache(app)
 
 
 @app.before_request
 def verify_token():
     """
-    This function distinguishes between requests that need authentication and  verifies if those who need to be
-    authenticated contain the authorization token in its headers or cookies and verifies
-    with keycloak if the token is valid. If the verification was successful, it updates the headers with
-    the user-info received from keycloak.
-    :return: 200 if authentication is not required, 200 if authentication is required and token is present and valid,
-    401 otherwise.
+    This function distinguishes between requests that need authorization
+    and verifies if those who need to be authorized contain the access
+    token in its headers or cookies. If the token verification and user
+    authorization was successful, it updates the headers, adding a
+    user-info string.
+    :return: HTTP response. 200 if path is not protected. 200 if path is
+    protected and meets the following criteria: 1) request contains an
+    Authorization header or cookie with bearer token. 2) The access
+    token has a valid signature (not expired or invalid). 3) The user
+    belongs to the appropriate group required to access the protected
+    path. 401 if token is invalid or not present. 403 if token is
+    present and valid but the user does not belong to the appropriate
+    groups for the protected path.
     """
-    # check if the url is for the /healthz route, in the future we might need to check what is the actual rule
+    # check if the url is for the /healthz route, in the future we might
+    # need to check what is the actual rule
     if request.url_rule is not None:
         return healthz()
 
-    # check if the url contains a keyword for the IN-CORE services
-    if '/dfr3' in request.url or '/data' in request.url or '/hazard' in request.url \
-            or '/space' in request.url or '/service' in request.url:
-        headers = {}
-        if request.headers.get('Authorization') is not None:
-            headers['Authorization'] = unquote_plus(request.headers['Authorization'])
-        elif request.cookies.get('Authorization') is not None:
-            headers['Authorization'] = unquote_plus(request.cookies['Authorization'])
-        else:
-            response = make_response('Unauthorized', 401)
-            return response
-
-        response = get_user_info_from_cache(headers['Authorization'])
-        if response is not None:
-            return response
-
-        # if token not in cache, return response from keycloak
-        return get_user_info_from_keycloak(headers)
-
-    else:
+    # retrieve resource from url
+    path = request.url[len(request.url_root):]
+    try:
+        resource = path.split('/')[0]
+    except IndexError:
         return Response(status=200)
+
+    if resource not in app.config["PROTECTED_RESOURCES"]:
+        return Response(status=200)
+
+    # retrieve access token from header or cookies
+    headers = {}
+    if request.headers.get('Authorization') is not None:
+        headers['Authorization'] = unquote_plus(
+            request.headers['Authorization'])
+    elif request.cookies.get('Authorization') is not None:
+        headers['Authorization'] = unquote_plus(
+            request.cookies['Authorization'])
+    else:
+        response = make_response('Unauthorized', 401)
+        return response
+    try:
+        access_token = headers['Authorization'].split(" ")[1]
+    except IndexError:
+        return make_response('Invalid token', 401)
+    public_key = f"-----BEGIN PUBLIC KEY-----\n" \
+        f"{str(os.environ.get('KEYCLOAK_PUBLIC_KEY'))}" \
+        f"\n-----END PUBLIC KEY-----"
+    audience = os.environ.get('KEYCLOAK_AUDIENCE', None)
+    if audience == "":
+        audience = None
+
+    # decode token for validating its signature
+    try:
+        access_token = jwt.decode(access_token, public_key,
+                                  audience = audience)
+    except ExpiredSignatureError:
+        return make_response(
+            'JWT Expired Signature Error: token signature has expired', 401)
+    except JWTClaimsError:
+        return make_response('JWT Claims Error: token signature is invalid',
+                             401)
+    except JWTError:
+        return make_response('JWT Error: token signature is invalid', 401)
+    except Exception:
+        return make_response('JWT Error: invalid token', 401)
+
+    # retrieve the groups the user belongs to from access token
+    user_groups = access_token.get("groups", [])
+    if "roles" in access_token:
+        user_roles = access_token["roles"]
+    elif "realm_access" in access_token:
+        user_roles = access_token["realm_access"].get("roles", [])
+    else:
+        user_roles = []
+
+    # get all resources the user can access to, based on the groups
+    # the user belongs to
+    user_accessible_resources = []
+    if "GROUPS" in app.config:
+        for group in user_groups:
+            if group in app.config["GROUPS"]:
+                user_accessible_resources.extend(app.config["GROUPS"][group])
+    if "ROLES" in app.config:
+        for role in user_roles:
+            if role in app.config["ROLES"]:
+                user_accessible_resources.extend(app.config["ROLES"][role])
+
+    if resource not in user_accessible_resources:
+        return make_response("access denied", 403)
+
+    # form user-info and add it to the response headers
+    user_info = {"preferred_username": access_token["preferred_username"]}
+    response = Response(status=200)
+    response.headers['x-auth-userinfo'] = json.dumps(user_info)
+    response.headers['Authorization'] = headers['Authorization']
+    return response
 
 
 @app.route("/healthz", methods=["GET"])
 def healthz():
-    return Response("Healthz OK", 200)
+    return Response("OK", 200)
 
 
 @app.before_first_request
@@ -66,43 +125,6 @@ def setup():
     if not app.debug:
         app.logger.addHandler(logging.StreamHandler())
         app.logger.setLevel(logging.INFO)
-
-
-def get_user_info_from_cache(token: str):
-    """
-    Try to get user info from cache
-    :param token: bearer access token
-    :return: 200 with updated headers if token is in the cache, None otherwise
-    """
-    user_info = cache.get(token)
-    if user_info is None:
-        return None
-
-    response = Response(status=200)
-    response.headers['x-auth-userinfo'] = user_info
-    response.headers['Authorization'] = token
-    # for testing purposes
-    # app.logger.info(user_info)
-    return response
-
-
-def get_user_info_from_keycloak(headers: dict):
-    """
-    Try to get user info from keycloak
-    :param headers: dictionary containing Authorization bearer access token
-    :return: HTTP response 200 with updated headers containing user-info, HTTP response 401 otherwise
-    """
-    url = cfg.KEYCLOAK_URL
-    r = requests.get(url, headers=headers)
-    if r.status_code == 200:
-        response = Response(status=200)
-        cache.set(headers['Authorization'], r.text)
-        response.headers['x-auth-userinfo'] = r.text
-        response.headers['Authorization'] = headers['Authorization']
-        app.logger.info(r.text)
-        return response
-    return Response(status=401)
-
 
 # for testing locally
 # if __name__ == "__main__":